requirements.txt
torch>=1.9.0
transformers>=4.21.0
numpy>=1.21.0
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
python-dotenv>=0.19.0
